# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/users": {
      "post": {
        "description": "Create a new user in the organization",
        "operationId": "UsersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "User details",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "User email address",
                    "example": "john.doe@example.com"
                  },
                  "name": {
                    "type": "string",
                    "description": "User full name",
                    "example": "John Doe"
                  },
                  "role": {
                    "type": "string",
                    "description": "User role",
                    "enum": [
                      "admin",
                      "member",
                      "viewer"
                    ],
                    "example": "member"
                  },
                  "profile": {
                    "type": "object",
                    "description": "User profile information",
                    "example": {
                      "department": "Engineering",
                      "title": "Software Engineer"
                    }
                  }
                },
                "required": [
                  "email",
                  "name"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "example": "user_abc123"
                    },
                    "email": {
                      "type": "string",
                      "example": "john.doe@example.com"
                    },
                    "name": {
                      "type": "string",
                      "example": "John Doe"
                    },
                    "role": {
                      "type": "string",
                      "example": "member"
                    },
                    "profile": {
                      "type": "object",
                      "example": {
                        "department": "Engineering"
                      }
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid user data"
          },
          "409": {
            "description": "User with this email already exists"
          }
        },
        "security": [
          {
            "x-user-id": []
          },
          {
            "x-org-id": []
          }
        ],
        "summary": "Create User",
        "tags": [
          "Users"
        ]
      },
      "get": {
        "description": "Get all users in the organization",
        "operationId": "UsersController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successfully retrieved users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "example": "user_abc123"
                      },
                      "email": {
                        "type": "string",
                        "example": "john.doe@example.com"
                      },
                      "name": {
                        "type": "string",
                        "example": "John Doe"
                      },
                      "role": {
                        "type": "string",
                        "example": "member"
                      },
                      "profile": {
                        "type": "object",
                        "example": {
                          "department": "Engineering"
                        }
                      },
                      "createdAt": {
                        "type": "string",
                        "format": "date-time"
                      },
                      "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "x-user-id": []
          },
          {
            "x-org-id": []
          }
        ],
        "summary": "List Users",
        "tags": [
          "Users"
        ]
      }
    },
    "/users/{id}": {
      "get": {
        "description": "Get a specific user by ID",
        "operationId": "UsersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User ID",
            "schema": {
              "example": "user_abc123",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved user",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "example": "user_abc123"
                    },
                    "email": {
                      "type": "string",
                      "example": "john.doe@example.com"
                    },
                    "name": {
                      "type": "string",
                      "example": "John Doe"
                    },
                    "role": {
                      "type": "string",
                      "example": "member"
                    },
                    "profile": {
                      "type": "object",
                      "example": {
                        "department": "Engineering"
                      }
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "x-user-id": []
          },
          {
            "x-org-id": []
          }
        ],
        "summary": "Get User",
        "tags": [
          "Users"
        ]
      },
      "put": {
        "description": "Update user details",
        "operationId": "UsersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User ID",
            "schema": {
              "example": "user_abc123",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Updated user details",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "User email address",
                    "example": "john.doe@example.com"
                  },
                  "name": {
                    "type": "string",
                    "description": "User full name",
                    "example": "John Doe"
                  },
                  "role": {
                    "type": "string",
                    "description": "User role",
                    "enum": [
                      "admin",
                      "member",
                      "viewer"
                    ],
                    "example": "member"
                  },
                  "profile": {
                    "type": "object",
                    "description": "User profile information",
                    "example": {
                      "department": "Engineering",
                      "title": "Senior Software Engineer"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User updated successfully"
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "x-user-id": []
          },
          {
            "x-org-id": []
          }
        ],
        "summary": "Update User",
        "tags": [
          "Users"
        ]
      },
      "delete": {
        "description": "Delete a user permanently",
        "operationId": "UsersController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User ID",
            "schema": {
              "example": "user_abc123",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "User deleted successfully"
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "x-user-id": []
          },
          {
            "x-org-id": []
          }
        ],
        "summary": "Delete User",
        "tags": [
          "Users"
        ]
      }
    },
...
  },
  "info": {
    "title": "Tolstoy API",
    "description": "Interactive API reference for Tolstoy workflow automation platform",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "https://tolstoy.getpullse.com",
      "description": "Production"
    },
    {
      "url": "http://localhost:3000",
      "description": "Development"
    }
  ],
  "components": {
    "securitySchemes": {
      "x-org-id": {
        "type": "apiKey",
        "in": "header",
        "name": "x-org-id",
        "description": "Organization ID for multi-tenant access"
      },
      "x-user-id": {
        "type": "apiKey",
        "in": "header",
        "name": "x-user-id",
        "description": "User ID for request context"
      }
    },
...
  }
}